backend/requirements.txt
alembic==1.14.0
flake8==6.1.0
black==24.3.0
fastapi==0.104.1
GeoAlchemy2==0.16.0
geojson-pydantic==1.1.2
shapely==2.0.1
geopandas==0.14.1
pandas==2.1.0
uvicorn[standard]==0.24.0
gunicorn==21.0.0
rasterio==1.3.10
numpy==1.26.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.1
psycopg2-binary==2.9.10
requests==2.31.0
aiohttp==3.8.6
tqdm==4.66.6